pyproject.toml
[tool.poetry]
name = "crm-backend"
version = "0.1.0"
description = "AI-enriched CRM backend with video/audio processing"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"
channels = "^4.0"
channels-redis = "^4.2"
daphne = "^4.1"
celery = "^5.3"
redis = "^5.0"
psycopg2-binary = "^2.9"
pgvector = "^0.2"
python-dotenv = "^1.0"
pillow = "^10.2"
numpy = "^1.26"
openai = "^1.12"
openai-whisper = "^20231117"
torch = "^2.2"
torchaudio = "^2.2"
face-recognition = "^1.3"
deepface = "^0.0.90"
transformers = "^4.37"
pydub = "^0.25"
scipy = "^1.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
black = "^24.1"
flake8 = "^7.0"
ipython = "^8.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
